+++ src/App.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+import React from "react";
+import logo from "./assets/logo.svg";
+import "./styles/App.css";
 
 function App() {
   return (
