+++ src/pages/Page.tsx
@@ -1,0 +1,1 @@
+export {};
